+++ controllers/contacts/changeContact.js
@@ -1,0 +1,21 @@
+const { HttpError } = require("../helpers");
+
+const Contact = require("../models");
+
+const changeContact = async (req, res, next) => {
+  try {
+    if (!Object.keys(req.body).length) {
+      return res.status(400).json({ message: "missing fields" });
+    }
+    const { id } = req.params;
+    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+    if (!result) {
+      throw HttpError(404, `Contact with ${id} not found`);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = changeContact;
