+++ controllers/contacts/deleteContact.js
@@ -1,0 +1,20 @@
+const { HttpError } = require("../helpers");
+
+const Contact = require("../models");
+
+const deleteContact = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const result = await Contact.findByIdAndDelete(id);
+    if (!result) {
+      throw HttpError(404, `Contact with ${id} not found`);
+    }
+    res.json({
+      message: "Delete success",
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = deleteContact;
