+++ controllers/contacts/getContactById.js
@@ -1,0 +1,18 @@
+const { HttpError } = require("../helpers");
+
+const Contact = require("../models");
+
+const getContactById = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const result = await Contact.findById(id);
+    if (!result) {
+      throw HttpError(404, `Contact with ${id} not found`);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getContactById;
