+++ controllers/contacts/getContacts.js
@@ -1,0 +1,12 @@
+const Contact = require("../models");
+
+const getContacts = async (req, res, next) => {
+  try {
+    const result = await Contact.find();
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = getContacts;
