--- controllers/contactsControllers.js
@@ -1,91 +1,0 @@
-const { HttpError } = require("../helpers");
-
-const Contact = require("../models");
-
-const getContacts = async (req, res, next) => {
-  try {
-    const result = await Contact.find();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const getContactById = async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await Contact.findById(id);
-    if (!result) {
-      throw HttpError(404, `Contact with ${id} not found`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const postContact = async (req, res, next) => {
-  try {
-    const result = await Contact.create(req.body);
-
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const deleteContact = async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await Contact.findByIdAndDelete(id);
-    if (!result) {
-      throw HttpError(404, `Contact with ${id} not found`);
-    }
-    res.json({
-      message: "Delete success",
-    });
-  } catch (error) {
-    next(error);
-  }
-};
-
-const changeContact = async (req, res, next) => {
-  try {
-    if (!Object.keys(req.body).length) {
-      return res.status(400).json({ message: "missing fields" });
-    }
-    const { id } = req.params;
-    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-    if (!result) {
-      throw HttpError(404, `Contact with ${id} not found`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateStatusContact = async (req, res, next) => {
-  try {
-    if (!Object.keys(req.body).length) {
-      return res.status(400).json({ message: "missing fields" });
-    }
-    const { id } = req.params;
-    const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
-    if (!result) {
-      throw HttpError(404, `Contact with ${id} not found`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-module.exports = {
-  getContacts,
-  getContactById,
-  postContact,
-  deleteContact,
-  changeContact,
-  updateStatusContact,
-};
